code decouple
